requirements.txt
streamlit==0.86.0
pandas==1.3.1
matplotlib
Pillow
